requirements.txt
torch
matplotlib
albumentations
Pillow
numpy
torchmetrics